# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QUEPON GGX Command System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pcs/{pcId}:
    get:
      operationId: getPc
      tags: [pcs]
      summary: Get a single PC detail
      parameters:
        - name: pcId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PC detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pc"
        "404":
          description: Not found
  /pcs/{pcId}/status:
    patch:
      operationId: updatePcStatus
      tags: [pcs]
      summary: Update PC status (admin)
      parameters:
        - name: pcId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PcStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pc"

  # --- QUEUE ---
...
  /promos:
    get:
      operationId: listPromos
      tags: [promos]
      summary: List all active promos
      responses:
        "200":
          description: Promos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promo"

    post:
      operationId: createPromo
      tags: [promos]
      summary: Create a promo (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promo"
  /promos/{promoId}:
    patch:
      operationId: updatePromo
      tags: [promos]
      summary: Update a promo (admin)
      parameters:
        - name: promoId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promo"

    delete:
      operationId: deletePromo
      tags: [promos]
      summary: Delete a promo (admin)
      parameters:
        - name: promoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # --- ANNOUNCEMENTS ---
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [promos]
      summary: List all active announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"

    post:
      operationId: createAnnouncement
      tags: [promos]
      summary: Create an announcement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{announcementId}:
    patch:
      operationId: updateAnnouncement
      tags: [promos]
      summary: Update an announcement (admin)
      parameters:
        - name: announcementId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

    delete:
      operationId: deleteAnnouncement
      tags: [promos]
      summary: Delete an announcement (admin)
      parameters:
        - name: announcementId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # --- FEEDBACK ---
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /players/{userId}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player detail (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /players/{userId}/status:
    patch:
      operationId: updatePlayerStatus
      tags: [players]
      summary: Update player status (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # --- WALLET ---
  /wallet/topup:
    post:
      operationId: topUpWallet
      tags: [wallet]
      summary: Top up a player's wallet (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopUpInput"
      responses:
        "200":
          description: Topped up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransaction"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, role, status, walletBalance, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [player, admin, superAdmin]
        status:
          type: string
          enum: [active, disabled, banned]
        walletBalance:
          type: number
        totalSpent:
          type: number
        totalPlayTimeMinutes:
          type: integer
        sessionCount:
          type: integer
        createdAt:
          type: string

    # --- PCs ---
    PcSpecs:
      type: object
      properties:
        cpu:
          type: string
        gpu:
          type: string
        ram:
          type: string
        monitor:
          type: string
    Pc:
      type: object
      required: [id, number, label, status, tier, createdAt]
      properties:
        id:
          type: string
        number:
          type: integer
        label:
          type: string
        status:
          type: string
          enum: [available, inUse, maintenance, reserved, cleaning, offline]
        tier:
          type: string
          enum: [standard, premium, vip]
        specs:
          $ref: "#/components/schemas/PcSpecs"
        currentSessionId:
          type: ["string", "null"]
        currentUserId:
          type: ["string", "null"]
        currentUsername:
          type: ["string", "null"]
        remainingSeconds:
          type: ["integer", "null"]
        location:
          type: string
        maintenanceNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    PcStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [available, inUse, maintenance, reserved, cleaning, offline]
        maintenanceNote:
          type: ["string", "null"]

    # --- QUEUE ---
...
    Promo:
      type: object
      required: [id, title, description, tag, isActive, displayPriority, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        tag:
          type: string
          enum: [discount, announcement, event, loyalty, maintenance, pricing]
        isActive:
          type: boolean
        displayPriority:
          type: integer
        startsAt:
          type: ["string", "null"]
        endsAt:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
    PromoInput:
      type: object
      required: [title, description, tag]
      properties:
        title:
          type: string
        description:
          type: string
        tag:
          type: string
        isActive:
          type: boolean
        displayPriority:
          type: integer
        startsAt:
          type: ["string", "null"]
        endsAt:
          type: ["string", "null"]
    PromoUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        tag:
          type: string
        isActive:
          type: boolean
        displayPriority:
          type: integer
        startsAt:
          type: ["string", "null"]
        endsAt:
          type: ["string", "null"]

    # --- ANNOUNCEMENTS ---
    Announcement:
      type: object
      required: [id, title, body, type, isActive, priority, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [general, maintenance, pricing, event, system]
        isActive:
          type: boolean
        priority:
          type: string
          enum: [low, normal, high, urgent]
        expiresAt:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, body, type]
      properties:
        title:
          type: string
        body:
          type: string
        type:
          type: string
        isActive:
          type: boolean
        priority:
          type: string
        expiresAt:
          type: ["string", "null"]
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        type:
          type: string
        isActive:
          type: boolean
        priority:
          type: string
        expiresAt:
          type: ["string", "null"]

    # --- FEEDBACK ---
...
    PlayerStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, disabled, banned]

    # --- WALLET ---
    WalletTransaction:
      type: object
      required: [id, userId, type, amount, previousBalance, newBalance, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: ["string", "null"]
        type:
          type: string
          enum: [topUp, sessionCharge, orderCharge, refund, adjustment]
        amount:
          type: number
        previousBalance:
          type: number
        newBalance:
          type: number
        paymentMethod:
          type: ["string", "null"]
        sessionId:
          type: ["string", "null"]
        orderId:
          type: ["string", "null"]
        processedBy:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
    TopUpInput:
      type: object
      required: [userId, amount, paymentMethod]
      properties:
        userId:
          type: string
        amount:
          type: number
          minimum: 1
        paymentMethod:
          type: string
          enum: [cash, gcash, maya, card, manualAdjustment]
        adminNote:
          type: ["string", "null"]

    # --- MENU ---
...