# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QUEPON GGX Command System API
 * OpenAPI spec version: 0.1.0
 */

export interface AnnouncementUpdate {
  title?: string;
  body?: string;
  type?: string;
  isActive?: boolean;
  priority?: string;
  /** @nullable */
  expiresAt?: string | null;
}