# lib/api-zod/src/generated/types/walletTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QUEPON GGX Command System API
 * OpenAPI spec version: 0.1.0
 */
import type { WalletTransactionType } from "./walletTransactionType";

export interface WalletTransaction {
  id: string;
  userId: string;
  /** @nullable */
  username?: string | null;
  type: WalletTransactionType;
  amount: number;
  previousBalance: number;
  newBalance: number;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  processedBy?: string | null;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
}